# openapi.yaml
openapi: 3.1.0
info:
  title: Order & Payment Microservices API
  version: 1.0.0
  description: |-
    REST APIs for the Clean Architecture-based Order and Payment microservices.
    The Order Service orchestrates payments and exposes idempotent order creation.
    The Payment Service enforces transaction limits, publishes payment.completed events,
    and surfaces the per-order status.
servers:
  - url: http://localhost:18080
    description: Order Service (local)
  - url: http://localhost:18081
    description: Payment Service (local)
paths:
...
  /payments:
    servers:
      - url: http://localhost:18081
        description: Payment Service (local)
    post:
      summary: Process a payment for an order
      operationId: processPayment
      tags:
        - payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Payment stored (Authorized or Declined)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      summary: List all payments
      operationId: listPayments
      tags:
        - payments
      responses:
        "200":
          description: Payments retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /payments/{order_id}:
    servers:
      - url: http://localhost:18081
        description: Payment Service (local)
    get:
      summary: Check payment status by order ID
      operationId: getPayment
      tags:
        - payments
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment details returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Payment not found for provided order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
...
  schemas:
...
    CreatePaymentRequest:
      type: object
      properties:
        order_id:
          type: string
        amount:
          type: integer
          format: int64
        customer_email:
          type: string
          format: email
      required:
        - order_id
        - amount
        - customer_email
...
    Payment:
      type: object
      properties:
        id:
          type: string
        order_id:
          type: string
        transaction_id:
          type: string
          nullable: true
        amount:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - Authorized
            - Declined
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string